+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### 4.8-SNAPSHOT
 #### Bugs
 
+* Fix #2023: Class RawCustomResourceOperationsImpl can't handle HTTP responses with empty body coming from the k8s 
+cluster (Jackson deserialization error was throwed). This kind of response can be returned after executing operations 
+like the delete of a custom resource. 
+
 #### Improvements
 
 * Fix #2012: osgi: Allow the ManagedKubernetesClient to consume an available OAuthTokenProvider